Added "Choose" Page permision.
This permission is intended to fully separate Sites on a multi-tenant Wagtail
installation. Not having "Choose" permission on a Page tree will prevent
the user from knowing that tree exists at all.

Eventually, everything that displays Pages in the
admin will be aware of this permission. And a similar permission will exist for
Collections as well.