Refactored the "actions" list render code to not require an API change.
The removal of all the "actions" buttons for required ancestor pages is now done
by avoiding the call to {% page_listing_buttons %} entirely. I realized that if
a user is not allowed to explore a page except by dint of it being a required
ancestor, they don't have permission to do *anything* to that page. So rendering
the "actions" list at all is pointless.

The button listing render code no longer runs at all for required ancestors,
obviating the need for an API change.