Prevent database error when calling permission_order.register on app ready
Fixes #12742

Previously, `permission_order.register` performed a database lookup for the content type. This is invalid if called from an app `ready` method as the documentation suggests, because this may run before the database has been initialised. Instead, `register` now queues up the arguments it receives, and the content type lookup is constructed lazily on first call to `get_content_type_order_lookup` (which happens when the group edit view is requested).