Remove unused setuptools cmdclass customisations
We no longer publish any NPM package from this repository. There is no
package.json in the "client" directory, and we do not include anything
in the "client" directory in the final package either.

The bdist_egg command is deprecated in favour of wheels. We don't
distribute eggs, so the custom version with the static assets check is
not relevant.